% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv_box_plot.R
\name{biv_box_plot}
\alias{biv_box_plot}
\title{A function to display or save box plots created on all numeric features.}
\usage{
biv_box_plot(
  dataset,
  classVar,
  order = NULL,
  colors = NULL,
  facet = NULL,
  loc = NULL
)
}
\arguments{
\item{dataset}{Name of the data frame object}

\item{classVar}{Name of the target feature}

\item{order}{A vector listing the target feature labels in the desired order. To use default order leave this parameter. Default value NULL.}

\item{colors}{A vector listing which color to use to represent which target feature label. To have the function pick color leave this parameter. Deafult value NULL.}

\item{loc}{A string with the directory where you want to save the plots. If no location is provided the plots will be created and displayed but not stored as image files.}
}
\value{
Returns three outputs:
1. If loc = NULL, returns plots and displays in the RStudio Plots window,
2. If loc has location provided, creates and saves the plots. Doesn't display in RStudio.
}
\description{
Creates (or saves as png if asked) boxplots for all numeric features. Split the box plots based on target feature categories.
}
\examples{
library(dplyr)
library(ggplot2)
biv_box_plot(dataset = iris,
    classVar = Species,
             order = c("virginica", "versicolor", "setosa"),
                        colors = c("virginica" = "#32a897", "versicolor" = "#328bab", "setosa" = "#8031ad"))


}
