% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hello}
\alias{hello}
\title{Test Function}
\usage{
hello(x)
}
\arguments{
\item{x}{A text string}
}
\value{
The output from \code{\link{print}}
\dontrun{
hello("Steve")
}
}
\description{
Test Function
}
\examples{
hello("John")


}
