% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fasteda.R
\name{theme_fasteda}
\alias{theme_fasteda}
\title{A color customizable theme}
\usage{
theme_fasteda(color = "#2F4F4F")
}
\arguments{
\item{color}{A hexacode starting with `#`.}
}
\value{
A customised theme to be used in ggplots
}
\description{
This is a customizable ggplot theme. Where a user can change the default background color
of ggplot with the color of their choice by providing the hexa-code of the color inside the function.
}
